Added a new property "use-credentials" to allow for authenticated requests.
The default behavior of ServiceWorker's "fetch" method is to set the "credentials"
property to "omit", which means that our lovely, HTTPS request to the server isn't
going to do us much good since it probably needs to be performed by an authenticated
user.

I added a "use-credentials" property that sets "credentials" to "include", defaulting
this new property to "false" for backward compatibility.

For more information on the fetch request, see:
1. http://www.html5rocks.com/en/tutorials/service-worker/introduction/
2. https://fetch.spec.whatwg.org/#fetch-method
3. https://fetch.spec.whatwg.org/#dom-request