Added a "focus-url" property.
The big problem that I ran into when working with "platinum-push-messaging" on
Android (as a "home screen" app) was that clicking the notification would not
behave as expected.

If the app was closed, then clicking on it opened the app to the desired URL
(the value of the message "url" field).  However, if the app was already up
or open in the background, then clicking on a notification simply refocused
the app.  This did not trigger the "platinum-push-messaging-click" event, and
it was pretty useless; my app needed the URL to change (it used hash routing).

The service-worker requirement for a 100% exact URL match, while conservative,
does not match with the philosophy of single-page apps.  Essentially, if my app
is open anywhere, then it should be focused and sent the click event, which my
app could then handle appropriately.