Notify when "supported" changes.
This was significant for me because I was loading the "platinum-push-messaging"
element within a "dom-if" template.  I was able to bind to "enabled", but my
app shows a separate message if the browser doesn't support push-messaging yet.

By adding "notify: true" to the "supported" property, it allowed me to bind
to the "platinum-push-messaging" element within a "dom-if".